Don't use shell quoting characters in release name.
While this doesn't solve unicode-releated problems with /etc/os-release
or /etc/fedora-release, for example, it does mean that we won't have
problems with parsing this through shell scripts, which we do quite
often.

This uses /Punctuation apostrophe/, U+2019, which is the preferred
unicode character for a displayed apostrophe, as opposed to /typewriter
apostrophe/, U=0027, which is also the shell quote character.

Signed-off-by: Peter Jones <pjones@redhat.com>
Reviewed-by: Adam Jackson <ajax@redhat.com>