Make release infomation work better with recent yum
Bug 1040607 describes the issue. Though currently in fedora-release
redhat-release doesn't have a version, I still added a provides
for system-release(release) since the default (in yum) is now to check that
first. And I removed the version from system-release in case someone
sets their yum config to check that instead of the default.