make $releasever return "rawhide" on Rawhide
This is a fairly significant change which will make $releasever (as understood
by dnf) return "rawhide" instead of a number (e.g. 35) for Rawhide users. The
purpose is to have a constant identifier, that developers and repo owners can
rely on. This will make Rawhide a true rolling release which won't need careful
adjustments every time we branch - once on Rawhide, you'll always stay on
Rawhide. This should also significantly simplify repo management, because repos
will no longer need a separate config for Rawhide which is different from stable
releases. The same config file containing "$releasever" can be used (instead of
hardcoding "rawhide" in URL), and it will resolve correctly to a "rawhide/"
directory on the repo server.

The overall hope is that this change should help users (system maintenance),
developers (simplifying automation) and repo owners (release differentiation no
longer needed). Please read the included link for a longer discussion.

This change introduces a new variable `is_rawhide`, which should be set to `1`
or `0`, depending on whether this release is Rawhide or not. It will then set
some additional variables automatically, to simplify specfile maintenance.

This change is made possible by dnf looking at package provides, when
determining $releasever, and `system-release(releasever)` is the first provides
it looks at. We don't use it in stable releases, but we can set the value on
Rawhide and it will have precedence.

This change will also need an accompanying change in fedora-repos, to be
committed and built together.

Related: https://pagure.io/releng/issue/7445