Add fedora-release-iot edition pieces
This adds the fedora-release-iot package for the Fedora IoT edition.
Note we don't add the option in the convert-to-edition script because
as IoT is purely rpm-ostree based we can't covert to IoT edition in
the traditional manner. We don't yet tweak presets but this is something
we intend to do moving forward.

Signed-off-by: Peter Robinson <pbrobinson@fedoraproject.org>