Drop dependencies
In F21-F29, the editions were seperate add-on packages and the
fedora-release package was the only one that provided system-release.

In F30, we've moved to a setup where fedora-release[-$EDITION] can
provide system-release. As an unexpected side-effect, this resulted
in -server and -workstation having an unresolvable dependency-chain,
because the edition-specific subpackages can no longer be ordered
late in the transaction.

Those dependencies were there solely for the ability to stop an
installation from reporting itself as Server or Workstation if
those packages were removed. (This implied a minimal set of
functionality that could be relied upon to be present for those
Editions). Since this was a nice-to-have feature rather than a
critical feature, we're going to drop those dependencies and treat
the Edition as an install-time declaration. These will still be
installed by default on the relevant Editions in their comps.xml
definition.

The glib-compile-schemas execution is now handled by a filetrigger
in the glib2 package.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>