Use the icon logo for `LOGO` in os-release(5)
In the initial addition of LOGO to os-release(5), the
wrong logo is used in here. It is supposed to be the "square" logo
icon that is used to represent the distribution in various contexts,
but instead we used the full logo with the wordmark, which messes
things up for applications trying to represent the distribution
properly.

This fixes it to request 'fedora-logo-icon', which is our "square"
logo that has only the Fedora logo without the wordmark.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>