Include empty VERSION_CODENAME= field in os-release
/etc/os-release now supports a VERSION_CODENAME field that should
contain the codename of a release if any exists. Fedora does not
use a codename, so it should be set to the empty string.

This is because some tools (such as https://github.com/nir0s/distro
for Python) attempt to interpret first the VERSION_CODENAME and
then subsequently fall back to parsing the VERSION field for a
codename within parentheses. This means that the tools will return
a different value of codename depending on which Edition was
originall installed. Instead, we should ensure that it returns no
codename at all.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>