Enable kata-osbuilder-generate.service by default
kata is an OCI container runtime that essentially wraps a qemu/kvm
VM around around a traditional container image. Inside the VM runs
a small appliance OS, containing the minimally needed kernel modules,
systemd, the kata-agent, and not much else. This appliance is generated
from host content, mostly using dracut.

kata-osbuilder-generate.service is a oneshot startup service that
generates the appliance for the boot kernel, and then exits. It's
only installed when kata is installed.

Signed-off-by: Cole Robinson <crobinso@redhat.com>