Enable rpm-ostree repo-refresh for all active local users
gnome-software attempts to refresh the rpm-ostree repository in order to check
if updates are available, in order to notify the user.  This happens
regularly in the bakground, including when the the settings for
automatic download and installation of updates are disabled.

The refresh requires the org.projectatomic.rpmostree1.repo-refresh
polkit action to be permitted, which is currently only the case for
users in the wheel group who are logged in locally.

The result is that logging in to a normal user account (ie: not in
'wheel') results in a polkit dialog immediately being presented, and
then at regular intervals thereafter.

Because a simple refresh is a fairly harmless operation, and because
knowing about the availability of updates is useful, allow it for all
users who are logged in locally.