Enable plocate timer by default
plocate is a (conflicting) replacement for mlocate, so if we enable the
latter, we should also enable the former.

The service does not alter other services, doesn't require any configuration,
should not fail, and does not use the network.

The same patch should be applied to f35, f34, and f33.