Statically enable rpm-ostree-countme timer
Commit 059e64f generalized the rpm-ostree count me unit activation using
the systemd preset configuration file.

This previous change incorrectly assumed that all rpm-ostree based
variants where using the presets as source of truth work enabled
services, but this is right now only true for Fedora CoreOS.

Work is in progress to update Silverblue/Kinoite and IoT to preset-all
by default. However this change will most likely be F36+ only.

Thus this change statically enables the timer unit for IoT, Silverblue &
Kinoite. Users can still opt out of counting as before by masking the
unit and this will not re-enable it for those that opted-out earlier.

See:
- https://pagure.io/workstation-ostree-config/pull-request/246
- https://github.com/coreos/fedora-coreos-config/blob/testing-devel/manifests/ignition-and-ostree.yaml#L39..L48
- https://github.com/coreos/fedora-coreos-config/pull/73
- https://github.com/coreos/fedora-coreos-config/pull/77
- https://github.com/coreos/fedora-coreos-config/pull/122
- https://github.com/coreos/rpm-ostree/issues/1803

Fixes: 059e64f 90-default.preset: Enable rpm-ostree count me by default