presets: Use a commmon desktop preset and spin specific ones
Split the 80-workstation.preset into a 81-desktop.preset with presets
common to all Fedora desktop variants and keeps the varaint specific
changes in separated preset files: 80-workstation.preset &
80-kde.preset.

This change is a NOP for Workstation and Silverblue.
This change disables sshd by defaut and switches cups to the socket unit
for the KDE Spin and Kinoite.

See: https://pagure.io/fedora-kde/SIG/issue/126